Finalizado Teste pepino Com FoFo Inicializado Teste pepino Com Gatos

+++ clinica-veterinaria-baade/src/test/java/net/baade/IntegratedTest.java
@@ -1,5 +1,7 @@
 package net.baade;
 
+import java.math.BigDecimal;
+
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -11,11 +13,11 @@ public class IntegratedTest {
 		recibo.setNomeCliente( "Dave" );
 		recibo.setNomeAnimal( "Fofo" );
 		// recibo.setValorAtendimento( new Float( 13.0 ) );
-		recibo.listaItemAdd( new Item( "Consulta de rotina", new Float( 10.0 ) ) );
-		recibo.listaItemAdd( new Item( "Vacinação contra raiva", new Float( 3.0 ) ) );
+		recibo.listaItemAdd( new Item( "Consulta de rotina", new BigDecimal( 10.0 ) ) );
+		recibo.listaItemAdd( new Item( "Vacinação contra raiva", new BigDecimal( 3.0 ) ) );
 		Assert.assertEquals( "Dave", recibo.getNomeCliente() );
 		Assert.assertEquals( "Fofo", recibo.getNomeAnimal() );
-		Assert.assertEquals( new Float( 13.0 ), recibo.getValorAtendimento() );
+		Assert.assertEquals( new BigDecimal( 13.0 ), recibo.getValorAtendimento() );
 		Assert.assertEquals( "Consulta de rotina", recibo.getItens().get( 0 ).getDescricao() );
 		Assert.assertEquals( "Vacinação contra raiva", recibo.getItens().get( 1 ).getDescricao() );
